More Ishmael patches: param check, int overflow, meaningless sanity check (added test here).

+++ src/java/voldemort/utils/Time.java
@@ -26,15 +26,15 @@ import java.util.Date;
  */
 public interface Time {
 
-    public final static int US_PER_MS = 1000;
-    public final static int NS_PER_US = 1000;
-    public final static int NS_PER_MS = US_PER_MS * NS_PER_US;
-    public final static int MS_PER_SECOND = 1000;
-    public final static int US_PER_SECOND = US_PER_MS * MS_PER_SECOND;
-    public final static int NS_PER_SECOND = NS_PER_US * US_PER_SECOND;
-    public final static int SECONDS_PER_HOUR = 60 * 60;
-    public final static int SECONDS_PER_DAY = 24 * SECONDS_PER_HOUR;
-    public final static int MS_PER_DAY = SECONDS_PER_DAY * MS_PER_SECOND;
+    public final static long US_PER_MS = 1000;
+    public final static long NS_PER_US = 1000;
+    public final static long NS_PER_MS = US_PER_MS * NS_PER_US;
+    public final static long MS_PER_SECOND = 1000;
+    public final static long US_PER_SECOND = US_PER_MS * MS_PER_SECOND;
+    public final static long NS_PER_SECOND = NS_PER_US * US_PER_SECOND;
+    public final static long SECONDS_PER_HOUR = 60 * 60;
+    public final static long SECONDS_PER_DAY = 24 * SECONDS_PER_HOUR;
+    public final static long MS_PER_DAY = SECONDS_PER_DAY * MS_PER_SECOND;
 
     public long getMilliseconds();
